Show an error message when trying to play yourself.
Trying to create and join a game from the same browser is incompatible with the cookie-based feature of being able to return to an ongoing game even after navigating away from the page. It would be impossible to distinguish who is who. As an alternative, the message suggests to play one side from an Incognito window