Fix scroll-to-bottom-on-render bug (again)
The move history automatically scrolled to the end at each clock tick.
I first fixed this bug in commit bfb7dc. The fix was to scroll only if moveHistory changed. The fix relied on the fact that clocks used to *not* be inside moveHistory. The new fix is to scroll only if the number of moves in move history changes, which is not affected by the clocks.

Also in this commit:
- Removed default react README