Improved the ELO id input field.
Before, it was not possible to erase the ELO id beyond the minimum length of 4. That made it awkward (but not impossible) to change the first 4 characters.
This is now changed so there are no restrictions on the minimum length of the ELO id in the input field.
If the user joins a game with an ELO id shorter than the minimum legnth, it is automatically replaced by a random ELO id.