For the history of recent games, fetch only game metadata instead of full games.
This optimization makes it so the backend needs to send only a bit of metadata per game instead of full games. This is fast enough that we can remove the incremental loading logic that we had previously.
In addition to cleaning up the code, this fixes the bug where new games did not appear immediately in the recent game list after being played (the page needed to be reloaded).