AI: if the game is lost, choose the move that simply walks towards the goal.
This avoids the possible infinite loop in AI vs AI games where a player who is mathematically lost moves randomly and the player who is mathematically winning plays stalling moves that nonetheless are still evaluated as winning.
The previous fix was to make the losing AI "resign", but this approach allows games to be fully played out.