Play vs dumb AI on the website.
This makes it possible to play against a dumb AI that simply walks along the shortest path towards the goal.
The AI needs to be improved before it is even a trivial challenge. (Not to be confused with the experimental C++ AI in the AI/ folder, which is not integrated with the website.)