AI: Add draws by the one-move-rule.
Before, the AI would ignore the one-move-rule (which says that the game is a draw if the player who moved first reaches the goal but the other player is within distance 2 of their goal).
Now the AI tries to avoid drawing by the one-move-rule if it can win.

Also: the evaluation for winning and losing moves has been tweaked so that: (a) winning moves that win in fewer moves are preferred, and (b) losing moves that lose in more moves are preferred. This changes the previous behavior where the choice between winning moves was arbitrary, and the choice between losing moves preferred ones that got closer to the goal.