AI: tentatively prune bad moves.
Moves consisting of 2 walls in the same 2-edge-connected component which block the two chosen edge-disjoint paths between the player to move and their goal) are likely to be bad AND expensive to check if they are legal (they require 2 distance computations).

When generating the list of legal moves,instead of verifying if moves of this type are legal or not before being generated, they are generated anyway but with a score that indicates that these moves may actually be illegal. Only when it is time to explore one of those moves, it is verified that it is legal. The point of this is that bad moves are more likely to be pruned by alpha-beta pruning, so there are good chances that we will be able to skip the legality check entirely.

As a quick experiment: calculating a move on a 4x4 board with depth 6 would have required about 313M graph traversals. However, 49M of those were avoided by the above optimization. Of those, only 4.6M of them were eventually explored (resulting in 2 * 4.6M extra graph traversals).