AI: AI is now a C++ template parametrized by board dimensions, and all AI code moved to headers.
This switch enables having multiple board dimensions at the same time. This is useful for runtime benchmarks and tests, where we want to test and benchmark a variety of scenarios without having to recompile.

As a (more or less) consequence of this switch, the entire code base has been moved to headers:

Having the board dimensions as compile time constants allows some optimizations, so, in order to be able to have multiple types, the Graph class is turned into a template (and moved to the header). Since Graph is a template, so must be Situation and Negamax, and likewise for most of the rest of code. Given this, it was simpler and faster to just move everything to headers, even things that didn't need to (faster both in compile time, since linking is slow, and in terms of allowing compiler optimizations, since cross-compile unit optimizations are harder to make).

Other things done in this commit:
- Removed the Mover base class and its subclass Walker, while its other two subclasses Human, and Negamaxer, become non-polymorphic classes. The point of the Mover class was to be able to easily switch between different AI algorithms, but now it is clear that there will be a single AI, Negamax, that will evolve over time.
- Simplified the logic for playing an interactive game and moved it from main.cc to to interactive_game.h.
- Moved PrintBoardWithEdgeIndices from board_io.{h,cc} to Situation.h, removing the need for the board_io files.
- Removed the LTO (Link time optimization) flag, which slowed down compilation and is not necessary anymore since everything is in one compile unit (I think?).
- Since anonymous namespaces in headers are not recommended, things that used to be in anonymous namespaces in .cc files are now private methods of (new) wrapper classes in header files.
- Bunch of other smaller cleanups (and things that got messier) due to the refactor.