Turns Graph and Situation into PODs.
In particular, the constructors are replaced with factory functions.

I did this change because I have a hunch that PODs are very large arrays of PODs are faster to initialize (but I am not entirely sure).

This is in preparation of the fixed-sized TT (transposition table). The TT will be a chunk of memory in the heap in the order of GBs. Each entry will have size 12B to 40B, depending on the board dimensions. That means that there will be between 89M and 26M entries per GB. So I am trying to avoid millions of calls to Situation and Graph constructors.