Frontend: Update type packages.
Now, things are as up to date as they can be.

This is the output of npm outdated:

Package                      Current   Wanted   Latest  Location                                  Depended by
@testing-library/jest-dom      4.2.4    4.2.4   5.16.5  node_modules/@testing-library/jest-dom    frontend
@testing-library/react         9.5.0    9.5.0   13.4.0  node_modules/@testing-library/react       frontend
@testing-library/user-event    7.2.1    7.2.1   14.4.3  node_modules/@testing-library/user-event  frontend
@types/jest                  26.0.24  26.0.24   29.4.0  node_modules/@types/jest                  frontend
@types/react                 17.0.53  17.0.53  18.0.28  node_modules/@types/react                 frontend
@types/react-dom             17.0.18  17.0.18  18.0.10  node_modules/@types/react-dom             frontend
react                         17.0.2   17.0.2   18.2.0  node_modules/react                        frontend
react-dom                     17.0.2   17.0.2   18.2.0  node_modules/react-dom                    frontend

Unfortunately, React cannot be updated to 18 because of the dependency
on react-materialize, which seems to not be compatible. In the future,
we should remove that dependency and update React.

react-dom, @types/react, and @types/react-dom are all aligned with React 17.

The testing libraries I do not care if they are outdated, as there are no tests yet.