Include the em++ output for the AI.
Until now, we were building the code locally and uploading it to the
gh-pages branch, and it was published from there. From now on, we will
use vercel and vercel will be fetching the code from the master branch
and comiling it. Thus, it needs to be able to see the em++ output (the
ai.mjs file).