Big refactor to start using client-side routing.
Before, everything happened in the https://wallwars.net URL. Now, this
URL corresponds to the lobby, whereas:
- https://wallwars.net/game/:gameId corresponds to a game that can be joined with the given ID.
- https://wallwars.net/game/offline corresponds to a game that can be played offline.
- https://wallwars.net/game/computer corresponds to a game that can be played against the computer.
- https://wallwars.net/puzzle/:puzzleId corresponds to a puzzle with the given ID.

This required some big refactors. In particular: Before, the game page
(GamePage) was embedded into the lobby (LobbyPage). Now, the two are
siblings and children of the App component. This required moving a ton
of state from LobbyPage to App.

Also in this commit:
- Reduced the memory used by webassembly memory from 2gb to 600mb. (The
transposition table size was set to 512mb, but wasm was allocating 2gb.
600mb is enough.)
- Moved the socket out of the component hierarchy, so now it does not need
to be passed around as a prop.
- In the header, do not show "Code" for games that do not have a join
code.

There are still some minor issues to iron out in follow up commits.