Change the management of id tokens on the server.
High level overview:

Each socket connection now has an idToken field that
identifies the (human) user associated with this socket connection.
It is a persistent and unique id for each user (it is the key in the
player DB table). It is empty for non-logged in users (guests).

All connections start as guests. The idToken is set when a client sends the
logInOrSignUp message, which includes their idToken. Since a log off button
is not implemented yet, once the client logs in, that socket connection will
always be associated with that user.

The idToken should not be confused with socket.id, which identifies the
connection itself and is used to route messages to the appropriate client.
The socket.id is never empty, even for guest users. Events like refreshing
the browser create a new socket connection/id.

Id tokens allow things like continuing a game from a different device.
Changing devices creates a new socket connection/id, but the game can still
be identified by the idToken (continuing from a different connection is not
possible for guests).

One of the changes is that the client only sends the idToken once, when
calling logInOrSignUp. It is no longer necessary to pass it in multiple
messages, such as changeName, getSolvedPuzzles, or returnToOngoingGame.