Fix bug in the name change dialog.
Before, if a player tried to press enter to confirm the new name in the dialog, they would get an unexpected effect: a new game would be created.
That is because the lobby page had registered a special handler for enter as a shortcut for the "create game" button.
This shortcut has been removed in this commit to avoid this situation.